+++ lib.php
@@ -56,9 +56,25 @@ function tool_moodlebox_before_footer() {
             get_config('tool_moodlebox', 'buttonsinfooter')) {
 
         $thisplugindir = $CFG->dirroot . '/admin/tool/moodlebox/';
+
+        // Display date and time setting buttons
+        $datetimetriggerfilename = $thisplugindir . '.set-server-datetime';
+        $datetimesetform = new datetimeset_form();
+
+        if ($data = $datetimesetform->get_data()) {
+            if (!empty($data->submitbutton)) {
+                $datecommand = "date +%s -s @$data->currentdatetime";
+                file_put_contents($datetimetriggerfilename, "#!/bin/sh\n" . $datecommand . "\nexit 0\n");
+                \core\notification::warning(get_string('datetimemessage', 'tool_moodlebox'));
+            }
+        }
+
+        $output = html_writer::empty_tag("hr", array('id' => 'datetimesetbuttonsspacer'));
+        $output .= html_writer::div($datetimesetform->render(), "", array('id' => 'datetimesetbuttons'));
+
+        // Display restart and shutdown buttons
         $reboottriggerfilename = $thisplugindir . '.reboot-server';
         $shutdowntriggerfilename = $thisplugindir . '.shutdown-server';
-
         $restartshutdownform = new restartshutdown_form(null, null, 'post', '', array('id' => 'formrestartstop'));
 
         if ($data = $restartshutdownform->get_data()) {
@@ -72,7 +88,7 @@ function tool_moodlebox_before_footer() {
             }
         }
 
-        $output = html_writer::empty_tag("hr", array('id' => 'footerbuttonsspacer'));
+        $output .= html_writer::empty_tag("hr", array('id' => 'footerbuttonsspacer'));
         $output .= html_writer::div($restartshutdownform->render(), "", array('id' => 'footerbuttons'));
 
         return $output;